requirements.txt
# requirements.txt
onnxruntime
silero
pydub
tqdm
pandas
numpy<2.0
PyYAML
requests